+++ src/class/config/class-get-options.php
@@ -167,18 +167,18 @@ class Get_Options extends Settings {
 	 * Build the URLs for popups
 	 *
 	 * @since 4.4 method added
-	 * @param 'movies'|'people'|'movies_search' $column Type of URL we want to get
+	 * @param 'movies'|'people'|'movies_search' $type_url Type of URL we want to get
 	 * @param string $domain_url OPTIONAL: Full URL of the domain, usually passed with site_url()
 	 * @return string
 	 */
-	public static function get_popup_url( string $column, $domain_url = '' ): string {
+	public static function get_popup_url( string $type_url, $domain_url = '' ): string {
 		$imdb_admin_option = get_option( self::get_admin_tablename() );
 		$url = [
-			'movies'        => $domain_url . $imdb_admin_option['imdburlpopups'] . parent::URL_BIT_POPUPS_MOVIES,
-			'people'        => $domain_url . $imdb_admin_option['imdburlpopups'] . parent::URL_BIT_POPUPS_PEOPLE,
-			'movies_search' => $domain_url . $imdb_admin_option['imdburlpopups'] . parent::URL_BIT_POPUPS_MOVIES_SEARCH,
+			'movies'        => $domain_url . $imdb_admin_option['imdburlpopups'] . parent::URL_BIT_POPUPS_FILM . '/',
+			'people'        => $domain_url . $imdb_admin_option['imdburlpopups'] . parent::URL_BIT_POPUPS_PERSON . '/',
+			'movies_search' => $domain_url . $imdb_admin_option['imdburlpopups'] . parent::URL_BIT_POPUPS_MOVIE_SEARCH . '/',
 		];
-		return $url[ $column ];
+		return $url[ $type_url ];
 	}
 }